Added device connection parameters.
The device has an comms_initialized variable that responds to a
constantly polling UNLOCK command. If the device is disconnected then
the device attempts to reconnect. This caused some issues when trying
to test disconnection tests. The device_available parameter instead is
an "uber connection" checker that will stop the device from being
reinitialized.